requirements.txt
asgiref==3.11.0
diff-match-patch==20241021
Django==6.0
django-filter==25.2
django-import-export==4.3.14
django-sslserver==0.22
django-unfold==0.74.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.95.0
drf-yasg==1.21.11
environs==14.5.0
gunicorn==23.0.0
inflection==0.5.1
marshmallow==4.1.1
packaging==25.0
pillow==12.0.0
PyJWT==2.10.1
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.5
tablib==3.9.0
uritemplate==4.2.0
